Move DB and http client to services to make ppm testable
